First prototype with flow information about institutions. Fixed styles and bugs of histogram table
Fixed the way the line chart interaction is woking to something cleaner
with a vertical bar browsing the data. Made histograms less wide is
less important to have wide space is more important for the line chart
and the bandwidth distribution.

Also the first initial prototype with flow data is set up. There is an
option on the menu to get information among the main institutions ,
still working on progress because we are shaping the form of the data
in the database.

Multiple styles changes to set up the table properly.