Better constructed Sankey. Fighting with Chord diagram formatting. Max, min, avg for flow showed input duplicated as input output.
ChordTest - example of chord diagram
dataparser- changes on the parsing of flow for max, min, avg bandwidth
trying to get data in both directions. (still not done right, so I
still duplicate the data of the input as output too)
Changed logo in dashboard with the new logo
readme.json file was the initial file I began playing to do the
chordTest.html it can be deleted.